+++ Packages/BrandonUtils/Testing/MultipleAsserter.cs
@@ -20,17 +20,20 @@ namespace BrandonUtils.Testing {
 
         private TActual GetActual() => Actual.Value;
 
-        [CanBeNull]
-        public Func<string> HeadingSupplier { get; set; }
+        [CanBeNull] public Func<string> HeadingSupplier { get; set; }
 
         [ItemNotNull]
-        public IList<Action> Actions { get; } = new List<Action>();
+        internal IList<Action> Actions_AgainstAnything { get; } = new List<Action>();
 
         [ItemNotNull]
-        public IList<Action<TActual>> ActionsAgainstActual { get; } = new List<Action<TActual>>();
+        internal IList<Action<TActual>> Actions_AgainstActual { get; } = new List<Action<TActual>>();
 
         [ItemNotNull]
-        public IList<IResolveConstraint> Constraints { get; } = new List<IResolveConstraint>();
+        internal IList<IResolveConstraint> Constraints_AgainstActual { get; } = new List<IResolveConstraint>();
+
+        internal IList<(object, IResolveConstraint)> Constraints_AgainstAnything { get; } = new List<(object, IResolveConstraint)>();
+
+        internal IList<(Func<TActual, object>, IResolveConstraint)> Constraints_AgainstTransformation { get; } = new List<(Func<TActual, object>, IResolveConstraint)>();
 
         protected abstract Action<string>                      ActionOnFailure    { get; }
         protected abstract Action<TActual, IResolveConstraint> ConstraintResolver { get; }
@@ -45,59 +48,131 @@ namespace BrandonUtils.Testing {
 
         private TSelf Self => this as TSelf;
 
+        #region Builder
+
+        [MustUseReturnValue]
         public TSelf Against([CanBeNull] TActual actual) {
             Actual = actual;
             return Self;
         }
 
+        #region "And" Constraints
+
+        #region Actions_AgainstAnything
+
+        [MustUseReturnValue]
         public TSelf And([CanBeNull] Action action) {
-            Actions.AddNonNull(action);
+            Actions_AgainstAnything.AddNonNull(action);
             return Self;
         }
 
+        [MustUseReturnValue]
         public TSelf And([CanBeNull] [ItemCanBeNull] IEnumerable<Action> actions) {
-            Actions.AddNonNull(actions);
+            Actions_AgainstAnything.AddNonNull(actions);
             return Self;
         }
 
+        #endregion
+
+        #region Actions_AgainstActual
+
+        [MustUseReturnValue]
         public TSelf And([CanBeNull] Action<TActual> action) {
-            ActionsAgainstActual.AddNonNull(action);
+            Actions_AgainstActual.AddNonNull(action);
             return Self;
         }
 
+        [MustUseReturnValue]
+        public TSelf And([CanBeNull] [ItemCanBeNull] IEnumerable<Action<TActual>> actions) {
+            Actions_AgainstActual.AddNonNull(actions);
+            return Self;
+        }
+
+        #endregion
+
+        #region Constraints_AgainstActual
+
+        [MustUseReturnValue]
         public TSelf And([CanBeNull] IResolveConstraint constraint) {
-            Constraints.AppendNonNull(constraint);
+            Constraints_AgainstActual.AddNonNull(constraint);
             return Self;
         }
 
+        [MustUseReturnValue]
         public TSelf And([CanBeNull] [ItemCanBeNull] IEnumerable<IResolveConstraint> constraints) {
-            Constraints.AddNonNull(constraints);
+            Constraints_AgainstActual.AddNonNull(constraints);
             return Self;
         }
 
+        #endregion
+
+        #region Constraints_AgainstAnything
+
+        [MustUseReturnValue]
+        public TSelf And((object, IResolveConstraint) constraint) {
+            Constraints_AgainstAnything.AddNonNull(constraint);
+            return Self;
+        }
+
+        [MustUseReturnValue]
+        public TSelf And([CanBeNull] IEnumerable<(object, IResolveConstraint)> constraints) {
+            Constraints_AgainstAnything.AddNonNull(constraints);
+            return Self;
+        }
+
+        #endregion
+
+        #region Constraints_AgainstTransformation
+
+        [MustUseReturnValue]
+        public TSelf And((Func<TActual, object>, IResolveConstraint) constraint) {
+            Constraints_AgainstTransformation.AddNonNull(constraint);
+            return Self;
+        }
+
+        [MustUseReturnValue]
+        public TSelf And([CanBeNull] IEnumerable<(Func<TActual, object>, IResolveConstraint)> constraints) {
+            Constraints_AgainstTransformation.AddNonNull(constraints);
+            return Self;
+        }
+
+        #endregion
+
+        #endregion
+
+        #region WithHeading
+
+        [MustUseReturnValue]
         public TSelf WithHeading([NotNull] Func<string> headingSupplier) {
             HeadingSupplier = headingSupplier;
             return Self;
         }
 
+        [MustUseReturnValue]
         public TSelf WithHeading([CanBeNull] string heading) {
             HeadingSupplier = () => heading;
             return Self;
         }
 
+        #endregion
+
+        #endregion
+
+        #region Executing Test Assertions
+
         /// <summary>
         /// </summary>
         /// <remarks>
         /// This method returns <see cref="IAssertable"/> instead of <see cref="Assertable"/> to make it play nice with the typed <see cref="Assertable{TActual}"/>,
-        /// which would be returned by <see cref="TestConstraint"/>.
+        /// which would be returned by <see cref="Test_Constraint_AgainstActual"/>.
         /// </remarks>
         /// <param name="action"></param>
         /// <returns></returns>
-        private static IAssertable TestAction([NotNull] Action action) {
+        private static IAssertable Test_Action_AgainstAnything([NotNull] Action action) {
             return new Assertable(action);
         }
 
-        private IAssertable TestActionAgainstActual([NotNull] Action<TActual> action) {
+        private IAssertable Test_Action_AgainstActual([NotNull] Action<TActual> action) {
             if (Actual.IsEmpty()) {
                 throw ActualIsEmptyException($"Could not execute the {action.GetType().Prettify()} {action.Method.Name}");
             }
@@ -105,13 +180,34 @@ namespace BrandonUtils.Testing {
             return new Assertable<TActual>(Actual.Value, action);
         }
 
-        private IAssertable TestConstraint([NotNull] IResolveConstraint constraint) {
+        private IAssertable Test_Constraint_AgainstActual([NotNull] IResolveConstraint constraint) {
             if (Actual.IsEmpty()) {
                 throw ActualIsEmptyException(constraint);
             }
 
             return new Assertable<TActual>(Actual.Value, constraint, ConstraintResolver);
         }
+
+        private IAssertable Test_Constraint_AgainstAnything([CanBeNull] object actual, [NotNull] IResolveConstraint constraint) {
+            return new Assertable<object>(actual, constraint, new Action<object, IResolveConstraint>(ConstraintResolver));
+        }
+
+        private IAssertable Test_Constraint_AgainstAnything((object, IResolveConstraint) constraint) {
+            var (actual, resolveConstraint) = constraint;
+            return Test_Constraint_AgainstAnything(actual, resolveConstraint);
+        }
+
+        private IAssertable Test_Constraint_AgainstTransformation((Func<TActual, object>, IResolveConstraint) constraint) {
+            var (transformation, resolveConstraint) = constraint;
+            if (Actual.IsEmpty()) {
+                throw ActualIsEmptyException(resolveConstraint);
+            }
+
+            var transformed = Actual.Select(transformation.Invoke);
+            return Test_Constraint_AgainstAnything(transformed, resolveConstraint);
+        }
+
+        #endregion
 
         private InvalidOperationException ActualIsEmptyException(IResolveConstraint constraint) {
             return ActualIsEmptyException($"Could not convert the {constraint.GetType().Name} to an {nameof(Action)}");
@@ -122,17 +218,16 @@ namespace BrandonUtils.Testing {
         }
 
         private IEnumerable<IAssertable> TestEverything() {
-            var actionFailures              = Actions.Select(TestAction);
-            var actionAgainstActualFailures = ActionsAgainstActual.Select(TestActionAgainstActual);
-            var constraintFailures          = Constraints.Select(TestConstraint);
-            return actionFailures
-                   .Concat(actionAgainstActualFailures)
-                   .Concat(constraintFailures);
+            return Actions_AgainstAnything.Select(Test_Action_AgainstAnything)
+                                          .Concat(Actions_AgainstActual.Select(Test_Action_AgainstActual))
+                                          .Concat(Constraints_AgainstActual.Select(Test_Constraint_AgainstActual))
+                                          .Concat(Constraints_AgainstAnything.Select(Test_Constraint_AgainstAnything))
+                                          .Concat(Constraints_AgainstTransformation.Select(Test_Constraint_AgainstTransformation));
         }
 
         public void ShortCircuit(Exception shortCircuitException) {
             ShortCircuitException = shortCircuitException;
-            Execute();
+            Invoke();
         }
 
         #region formatting
@@ -168,7 +263,7 @@ namespace BrandonUtils.Testing {
 
         #endregion
 
-        public void Execute() {
+        public void Invoke() {
             var failures = TestEverything().ToList();
             if (failures.Any(it => it.Failed)) {
                 ActionOnFailure.Invoke(FormatMultipleAssertionMessage(failures));